greenlake_app/requirements.txt
fastapi
uvicorn
jinja2
requests
requests-oauthlib
python-multipart
python-dotenv
aiofiles
pydantic
PyYAML
pytz
pydantic-settings